+++ Player/Extensions/Error+Extensions.swift
@@ -11,39 +11,24 @@ import AVFoundation
 extension Error {
     internal var debugInfoString: String {
         if let expandedError = self as? ExpandedError {
-            return expandedError.domain + " \(expandedError.code): " + expandedError.message + " " + (expandedError.info ?? "")
-        }
-        if let nsError = self as? NSError {
-            var message = ""
-            message += "Code=\(nsError.code) \n "
-            message += "Domain=\(nsError.domain) \n "
-            if let value = nsError.userInfo[NSDebugDescriptionErrorKey] as? String {
-                message += "Message=\(value) \n "
+            var message = "[\(expandedError.code):" + expandedError.domain + "] \n"
+            if let underlyingError = expandedError.underlyingError {
+                message += underlyingError.debugInfoString
             }
             else {
-                message += "Message=\(nsError.debugDescription) \n "
-            }
-            
-            if let value = nsError.userInfo[NSURLErrorKey] as? URL {
-                message += "URL=\(value.absoluteString) \n"
-            }
-            
-            if let uError = nsError.userInfo[NSUnderlyingErrorKey] as? NSError {
-                message += "UnderlyingCode=\(uError.code) \n "
-                message += "UnderlyingDomain=\(uError.domain) \n "
-                if let value = uError.userInfo[NSDebugDescriptionErrorKey] as? String {
-                    message += "UnderlyingMessage=[\(value)] \n "
-                }
-                else {
-                    message += "UnderlyingMessage=[\(uError.debugDescription)] \n "
-                }
-                
-                if let value = uError.userInfo[NSURLErrorKey] as? URL {
-                    message += "UnderlyingURL=\(value.absoluteString) \n"
-                }
+                message += "[" + expandedError.message + " " + (expandedError.info ?? "") + "]"
             }
             return message
         }
-        return self.localizedDescription
+        else if let nsError = self as? NSError {
+            var message = "[\(nsError.code):\(nsError.domain)] \n "
+            message += "[\(nsError.debugDescription)] \n "
+            
+            if let uError = nsError.userInfo[NSUnderlyingErrorKey] as? NSError {
+                message += uError.debugInfoString
+            }
+            return message
+        }
+        return "[\(self.localizedDescription)] \n"
     }
 }